+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In progress
 
 - Rework move structure and enhance move sheet ([#62](https://github.com/ben/foundry-ironsworn/pull/62))
+- Update roll-bonus dialog to allow selecting stat for moves ([#63](https://github.com/ben/foundry-ironsworn/pull/63))
 
 ## 0.4.9